+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # 2020-02-27-17:28 - geaysw
 
+## Features
+- Added `!website` command to pull up VMS website URL
+
 ## Fixes
 - Google Sheets API token refresh now works
+- Calendar command now returns the correct link
 
 # 2020-02-27-16:47 - gearysw